Add-ADUsersToGroups.ps1
In this script, we first define the AD groups we want to add and the AD users we want to add to. We then use nested loops to iterate through each AD user and each AD group, and add the user to the group using the Add-ADGroupMember cmdlet.

Just so you know, you will need to modify the $groups and $users variables with the actual names of the AD groups and AD users you want to add. Additionally, you may need to adjust the Identity parameter of the Add-ADGroupMember cmdlet to match the specific name or distinguished name of your AD groups.