Create StartMbamClientUiOnRemote.ps1
This script prompts for the name of the remote computer, the username for the remote computer, and the password for the remote computer (which is entered securely as a secure string).

The Read-Host cmdlet is used to prompt the user for input. The -AsSecureString parameter is used to mask the password input.